Add the ability to use file handles, in addition to supplying a path.
If the file being accessed is not within the file system (for example,
if it is compressed into a zip or tar file) it cannot be opened from
the supplied path. This change adds the ability to supply a file handle,
and for the module to use that, rather than opening the file itself.